Double to float

+++ Pixel Arena/Entity.cpp
@@ -13,17 +13,17 @@ Entity::Entity()
 
 }
 
-bool Entity::InView(sf::Vector2f View, double ViewAngle,double fov)
+bool Entity::InView(sf::Vector2f View, float ViewAngle, float fov)
 {
 	bool ReturnValue = 0;
 
 	//Calculate the Vector between View and 
-	double dX = this->x - View.x;
-	double dY = -(this->y - View.y);
+	float dX = this->x - View.x;
+	float dY = -(this->y - View.y);
 
-	double AngletoEntity = LogManager::FixAngle(LogManager::RadtoDeg(std::atan2f(dY, dX)));
-	double RightView = LogManager::FixAngle(ViewAngle - fov / 2.0);
-	double LeftView = LogManager::FixAngle(ViewAngle + fov / 2.0);
+	float AngletoEntity = LogManager::FixAngle(LogManager::RadtoDeg(std::atan2f(dY, dX)));
+	float RightView = LogManager::FixAngle(ViewAngle - fov / 2.0);
+	float LeftView = LogManager::FixAngle(ViewAngle + fov / 2.0);
 
 
 	//Check if in View
@@ -47,11 +47,11 @@ bool Entity::InView(sf::Vector2f View, double ViewAngle,double fov)
 
 	//AngletoEntity vs entity view angle
 	if (ReturnValue) {
-		double radEntity = LogManager::DegtoRad(this->DegAngle);
-		double radRotEntity = 0;//rotated entity angle as reference vector
-		double radRotToEntity = LogManager::DegtoRad(AngletoEntity) - radEntity;
+		float radEntity = LogManager::DegtoRad(this->DegAngle);
+		float radRotEntity = 0;//rotated entity angle as reference vector
+		float radRotToEntity = LogManager::DegtoRad(AngletoEntity) - radEntity;
 		//skip flipping due to -y in sfml
-		double degAngleBetween = LogManager::RadtoDeg(atan2f(sin(radRotToEntity),cos(radRotToEntity)));
+		float degAngleBetween = LogManager::RadtoDeg(atan2f(sin(radRotToEntity),cos(radRotToEntity)));
 
 		int dPAVE = int(degAngleBetween + 180); // degPlayerAngleVsEntity
 
@@ -73,7 +73,7 @@ bool Entity::InView(sf::Vector2f View, double ViewAngle,double fov)
 	return ReturnValue;
 }
 
-void Entity::Update(double distance)
+void Entity::Update(float distance)
 {
 	InternalClock.Count();
 	this->distanceToPlayer = distance;
@@ -85,7 +85,7 @@ sf::Vector2f Entity::getPosition() const
 	return sf::Vector2f(x,y);
 }
 
-void Entity::drawEntity(double* Z_Buffer, sf::RenderTarget& target, sf::RenderStates states)
+void Entity::drawEntity(float* Z_Buffer, sf::RenderTarget& target, sf::RenderStates states)
 {
 }
