+++ Pixel Arena/Orc.h
@@ -7,13 +7,13 @@ public:
 private:
 	sf::RectangleShape Object;
 public:
-	void drawEntity(double* Z_Buffer,sf::RenderTarget& target, sf::RenderStates states = sf::RenderStates::Default);
+	void drawEntity(float* Z_Buffer,sf::RenderTarget& target, sf::RenderStates states = sf::RenderStates::Default);
 
 private:
 	virtual void draw(sf::RenderTarget& target, sf::RenderStates states = sf::RenderStates::Default) const  override;
 
 	int AnimSizeX;//on construction
 	int AnimSizeY;
-	void UpdateAnimation(sf::RenderTarget& target, double* Z_Buffer);
+	void UpdateAnimation(sf::RenderTarget& target, float* Z_Buffer);
 };
